[144]-some-rebuild-and-fixes
